Extract to mbox, vcalendar files
Extract all items in Calendar folder to an mbox file, also output
all text/calendar parts to a .vcalendar file as a series of VEVENT
elements contained within a single VCALENDAR element.

Initial tests have shown that it might be useful to exclude long-
past events, and/or to split the events across multiple files
before importing to Google Calendar.

Currently, no warning is generated when a message in the
Calendar folder does not contain a text/calendar part.