Saving skipped messages, \n between mbox messages
I addition to support for saving messages which failed to import,
add the ability to save skipped messages. This function is not
currently used as of this commit.

Also, include a '\n' after each saved message to avoid issues
with line endings which manifest by seeing a '^M' at the end of
each message when viewed in mutt.